+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ Unreleased
 * feat(Instrument.intraday_best_limits): new method
 * feat(Instrument.intraday_states): new method
 * feat(Instrument.intraday_trades): new method
-* feat(Instrument.intraday_thresholds): new method
+* feat(Instrument.static_thresholds): new method
 * feat(Instrument.holders_by_date): new method
 * feat(Instrument.client_type_history): new method